fucking signal herdoc

+++ parsing/handle_herdoc.c
@@ -6,7 +6,7 @@
 /*   By: khbouych <khbouych@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/25 14:34:45 by mel-kouc          #+#    #+#             */
-/*   Updated: 2023/08/09 14:48:15 by khbouych         ###   ########.fr       */
+/*   Updated: 2023/08/15 16:31:11 by khbouych         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,25 +36,37 @@ void	write_in_herdoc(t_token *ptr, t_parse *new_p, t_env *env)
 	delim = ptr->content;
 	buffer = NULL;
 	new_p->f_name = generate_name();
-	ft_signals();
-	while (1)
+	int pid;
+	pid = fork();
+	if (pid == 0)
 	{
-		str = readline("> ");
-		if (!str || !ft_strncmp(delim, str, ft_strlen(str)))
+		while (1)
 		{
-			new_p->fd_input = open(new_p->f_name, O_WRONLY | O_CREAT | O_TRUNC, 0644);
-			if (buffer)
-				write(new_p->fd_input, buffer, ft_strlen(buffer));
-			break ;
+			signal(SIGINT, SIG_DFL);
+			signal(SIGQUIT, SIG_DFL);
+			str = readline("> ");
+			if (!str || !ft_strncmp(delim, str, ft_strlen(str)))
+			{
+				new_p->fd_input = open(new_p->f_name, O_WRONLY | O_CREAT | O_TRUNC, 0644);
+				if (buffer)
+					write(new_p->fd_input, buffer, ft_strlen(buffer));
+				break ;
+			}
+			if (ft_strchr(str, '$') != -1 && ptr->flag != 1)
+				str = ft_expandhelp(str, env);
+			buffer = ft_strjoin(buffer, str);
+			buffer = ft_strjoin(buffer, "\n");
 		}
-		if (ft_strchr(str, '$') != -1 && ptr->flag != 1)
-			str = ft_expandhelp(str, env);
-		buffer = ft_strjoin(buffer, str);
-		buffer = ft_strjoin(buffer, "\n");
+		close(new_p->fd_input);
+		new_p->fd_input = open(new_p->f_name, O_RDONLY | O_TRUNC, 0644);
+		unlink(new_p->f_name);
+		free(new_p->f_name);
 	}
-	close(new_p->fd_input);
-	new_p->fd_input = open(new_p->f_name, O_RDONLY | O_TRUNC, 0644);
-	unlink(new_p->f_name);
+	else
+	{
+		waitpid(pid, NULL, 0);
+		ft_signals();
+	}
 }
 
 void	ft_searsh_herdoc(t_token *tmp, t_parse *new_p, t_env *env)
